+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 - Histogram metric is added and regular counter no longer has an input value [#22](https://github.com/zlikavac32/metco/pull/22)
 
+## Changed
+
+- Removed gauge is persisted until new cycle [#23](https://github.com/zlikavac32/metco/pull/23)
+
 ## 0.7.0 - 2025-10-03
 
 ### Added